Move some of the files to .internal package
