+++ AndroidSDKCore/src/main/java/com/leanplum/Leanplum.java
@@ -28,8 +28,8 @@ import android.text.TextUtils;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import com.leanplum.ActionContext.ContextualValues;
-import com.leanplum.actions.ActionDefinition;
-import com.leanplum.actions.ActionManagerDefinitionKt;
+import com.leanplum.actions.internal.ActionDefinition;
+import com.leanplum.actions.internal.ActionManagerDefinitionKt;
 import com.leanplum.callbacks.ActionCallback;
 import com.leanplum.callbacks.ForceContentUpdateCallback;
 import com.leanplum.callbacks.RegisterDeviceCallback;
