+++ AndroidSDKCore/src/main/java/com/leanplum/actions/internal/ActionManagerTriggering.kt
@@ -19,7 +19,7 @@
  * under the License.
  */
 
-package com.leanplum.actions
+package com.leanplum.actions.internal
 
 import com.leanplum.ActionContext
 import com.leanplum.internal.ActionManager
