+++ AndroidSDKCore/src/main/java/com/leanplum/internal/ActionManager.java
@@ -25,16 +25,15 @@ import android.content.Context;
 import android.content.SharedPreferences;
 
 import android.text.TextUtils;
-import androidx.annotation.NonNull;
 import androidx.annotation.VisibleForTesting;
 import com.leanplum.ActionContext.ContextualValues;
 import com.leanplum.Leanplum;
 import com.leanplum.LocationManager;
-import com.leanplum.actions.Action;
-import com.leanplum.actions.ActionManagerExecutionKt;
-import com.leanplum.actions.ActionQueue;
-import com.leanplum.actions.ActionScheduler;
-import com.leanplum.actions.Definitions;
+import com.leanplum.actions.internal.Action;
+import com.leanplum.actions.internal.ActionManagerExecutionKt;
+import com.leanplum.actions.internal.ActionQueue;
+import com.leanplum.actions.internal.ActionScheduler;
+import com.leanplum.actions.internal.Definitions;
 import com.leanplum.actions.MessageDisplayController;
 import com.leanplum.actions.MessageDisplayListener;
 import com.leanplum.internal.Constants.Defaults;
