+++ AndroidSDKCore/src/main/java/com/leanplum/internal/LeanplumInternal.java
@@ -31,9 +31,9 @@ import com.leanplum.Leanplum;
 import com.leanplum.LeanplumActivityHelper;
 import com.leanplum.LeanplumException;
 import com.leanplum.LeanplumLocationAccuracyType;
-import com.leanplum.actions.ActionManagerTriggeringKt;
-import com.leanplum.actions.ActionsTrigger;
-import com.leanplum.actions.Priority;
+import com.leanplum.actions.internal.ActionManagerTriggeringKt;
+import com.leanplum.actions.internal.ActionsTrigger;
+import com.leanplum.actions.internal.Priority;
 import com.leanplum.callbacks.StartCallback;
 import com.leanplum.internal.Request.RequestType;
 import com.leanplum.models.GeofenceEventType;
