+++ AndroidSDKCore/src/main/java/com/leanplum/messagetemplates/MessageTemplates.java
@@ -27,8 +27,8 @@ import androidx.annotation.Nullable;
 import com.leanplum.ActionArgs;
 import com.leanplum.ActionContext;
 import com.leanplum.Leanplum;
-import com.leanplum.actions.ActionDefinition;
-import com.leanplum.actions.ActionManagerDefinitionKt;
+import com.leanplum.actions.internal.ActionDefinition;
+import com.leanplum.actions.internal.ActionManagerDefinitionKt;
 import com.leanplum.callbacks.ActionCallback;
 import com.leanplum.internal.ActionManager;
 import com.leanplum.messagetemplates.actions.AlertMessage;
